% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acl_article.R
\name{acl_paper}
\alias{acl_paper}
\title{ACL submission format (PDF)}
\usage{
acl_paper(keep_tex = TRUE, includes = NULL, submission = TRUE)
}
\arguments{
\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\description{
Template for creating a ACL submission
}

