% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogsci_article.R
\name{cogsci_paper}
\alias{cogsci_paper}
\title{Cogsci submission format (PDF)}
\usage{
cogsci_paper(keep_tex = TRUE, includes = NULL)
}
\arguments{
\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\description{
Template for creating a CogSci submission
}

